refactor: extract feedback JSON parsing into shared utility
